python-reporter/requirements.txt


pymongo==4.8.0
python-dotenv==1.0.1
certifi==2024.7.4